requirements.txt
hydra-core==1.3.2
matplotlib==3.7.2
monai==1.2.0
numpy==1.25.1
pandas==2.0.3
SimpleITK==2.2.1
timm==0.9.2
torch==2.0.1
torchio==0.19.0
torchmetrics==1.0.1
torchvision==0.15.2
tqdm==4.65.0
wandb==0.15.8
